requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Pillow==10.0.1
Werkzeug==2.3.7
bcrypt==4.1.2
requests==2.31.0
